api/requirements.txt
dj-rest-auth==7.0.1
Django==5.2.8
django-allauth==65.13.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
psycopg2-binary==2.9.11
requests==2.32.5
whitenoise==6.11.0
django-environ==0.12.0
django-cors-headers==4.9.0
bleach==6.2.0
tinycss2==1.4.0
pypdf==5.2.0
Pillow==11.1.0
